Experimental: lower the chunk-claiming range to Short.MIN/MAX_VALUE, to allow declaring seemingly-overlapping area

+++ src/main/java/org/teacon/areacontrol/AreaControlCommand.java
@@ -134,8 +134,8 @@ public final class AreaControlCommand {
         final var src = context.getSource();
         final var claimer = src.getPlayerOrException();
         final var chunkPos = new ChunkPos(new BlockPos(src.getPosition()));
-        final var chunkStart = chunkPos.getBlockAt(0, Integer.MIN_VALUE, 0);
-        final var chunkEnd = chunkPos.getBlockAt(15, Integer.MAX_VALUE, 15);
+        final var chunkStart = chunkPos.getBlockAt(0, Short.MIN_VALUE, 0);
+        final var chunkEnd = chunkPos.getBlockAt(15, Short.MAX_VALUE, 15);
         final var range = new AABB(Vec3.atCenterOf(chunkStart), Vec3.atCenterOf(chunkEnd));
         if (!range.expandTowards(0.5, 0.5, 0.5).contains(claimer.position())) {
             src.sendFailure(new TranslatableComponent("area_control.error.outside_selection"));
@@ -165,44 +165,44 @@ public final class AreaControlCommand {
         if (xOffset > 0) {
             if (zOffset > 0) {
                 // Corner is at northwest
-                chunkStart = corner.getBlockAt(0, Integer.MIN_VALUE, 0);
-                chunkEnd = new ChunkPos(corner.x + xOffset - 1, corner.z + zOffset - 1).getBlockAt(15, Integer.MAX_VALUE, 15);
+                chunkStart = corner.getBlockAt(0, Short.MIN_VALUE, 0);
+                chunkEnd = new ChunkPos(corner.x + xOffset - 1, corner.z + zOffset - 1).getBlockAt(15, Short.MAX_VALUE, 15);
             } else if (zOffset == 0) {
                 // Corner is at northwest
-                chunkStart = corner.getBlockAt(0, Integer.MIN_VALUE, 0);
-                chunkEnd = new ChunkPos(corner.x + xOffset - 1, corner.z).getBlockAt(15, Integer.MAX_VALUE, 15);
+                chunkStart = corner.getBlockAt(0, Short.MIN_VALUE, 0);
+                chunkEnd = new ChunkPos(corner.x + xOffset - 1, corner.z).getBlockAt(15, Short.MAX_VALUE, 15);
             } else {
                 // Corner is at southwest
-                chunkStart = corner.getBlockAt(0, Integer.MIN_VALUE, 15);
-                chunkEnd = new ChunkPos(corner.x + xOffset - 1, corner.z + zOffset + 1).getBlockAt(15, Integer.MAX_VALUE, 0);
+                chunkStart = corner.getBlockAt(0, Short.MIN_VALUE, 15);
+                chunkEnd = new ChunkPos(corner.x + xOffset - 1, corner.z + zOffset + 1).getBlockAt(15, Short.MAX_VALUE, 0);
             }
         } else if (xOffset == 0) {
             if (zOffset > 0) {
                 // Corner is at northwest
-                chunkStart = corner.getBlockAt(0, Integer.MIN_VALUE, 0);
-                chunkEnd = new ChunkPos(corner.x, corner.z + zOffset - 1).getBlockAt(15, Integer.MAX_VALUE, 15);
+                chunkStart = corner.getBlockAt(0, Short.MIN_VALUE, 0);
+                chunkEnd = new ChunkPos(corner.x, corner.z + zOffset - 1).getBlockAt(15, Short.MAX_VALUE, 15);
             } else if (zOffset == 0) {
                 // Corner is at northwest, just one chunk
-                chunkStart = corner.getBlockAt(0, Integer.MIN_VALUE, 0);
-                chunkEnd = corner.getBlockAt(15, Integer.MAX_VALUE, 15);
+                chunkStart = corner.getBlockAt(0, Short.MIN_VALUE, 0);
+                chunkEnd = corner.getBlockAt(15, Short.MAX_VALUE, 15);
             } else {
                 // Corner is at southwest
-                chunkStart = corner.getBlockAt(0, Integer.MIN_VALUE, 15);
-                chunkEnd = new ChunkPos(corner.x, corner.z + zOffset + 1).getBlockAt(15, Integer.MAX_VALUE, 0);
+                chunkStart = corner.getBlockAt(0, Short.MIN_VALUE, 15);
+                chunkEnd = new ChunkPos(corner.x, corner.z + zOffset + 1).getBlockAt(15, Short.MAX_VALUE, 0);
             }
         } else {
             if (zOffset > 0) {
                 // Corner is at northeast
-                chunkStart = corner.getBlockAt(15, Integer.MIN_VALUE, 0);
-                chunkEnd = new ChunkPos(corner.x + xOffset + 1, corner.z + zOffset - 1).getBlockAt(0, Integer.MAX_VALUE, 15);
+                chunkStart = corner.getBlockAt(15, Short.MIN_VALUE, 0);
+                chunkEnd = new ChunkPos(corner.x + xOffset + 1, corner.z + zOffset - 1).getBlockAt(0, Short.MAX_VALUE, 15);
             } else if (zOffset == 0) {
                 // Corner is at northeast
-                chunkStart = corner.getBlockAt(15, Integer.MIN_VALUE, 0);
-                chunkEnd = new ChunkPos(corner.x + xOffset + 1, corner.z).getBlockAt(0, Integer.MAX_VALUE, 15);
+                chunkStart = corner.getBlockAt(15, Short.MIN_VALUE, 0);
+                chunkEnd = new ChunkPos(corner.x + xOffset + 1, corner.z).getBlockAt(0, Short.MAX_VALUE, 15);
             } else {
                 // Corner is southeast
-                chunkStart = corner.getBlockAt(15, Integer.MIN_VALUE, 15);
-                chunkEnd = new ChunkPos(corner.x + xOffset + 1, corner.z + zOffset + 1).getBlockAt(0, Integer.MAX_VALUE, 0);
+                chunkStart = corner.getBlockAt(15, Short.MIN_VALUE, 15);
+                chunkEnd = new ChunkPos(corner.x + xOffset + 1, corner.z + zOffset + 1).getBlockAt(0, Short.MAX_VALUE, 0);
             }
         }
         final var range = new AABB(Vec3.atCenterOf(chunkStart), Vec3.atCenterOf(chunkEnd));
